refactor: handle default caip network if namespace is not available
